Sync up the update of CircuitPython library commit #c6763dff
Remove redundant
Add flashinfo.py
