+++ circuitPython/lib/bteve/mcu_ft80x.py
@@ -147,61 +147,6 @@ class Mcu_FT80X():
         self.RAM_DL_SIZE                = 8*1024
         self.RAM_PAL_SIZE               = 1*1024
 
-        # Coprocessor related commands
-        self.CMD_APPEND                 = 0xFFFFFF1E
-        self.CMD_BGCOLOR                = 0xFFFFFF09
-        self.CMD_BITMAP_TRANSFORM       = 0xFFFFFF21
-        self.CMD_BUTTON                 = 0xFFFFFF0D
-        self.CMD_CALIBRATE              = 0xFFFFFF15
-        self.CMD_CLOCK                  = 0xFFFFFF14
-        self.CMD_COLDSTART              = 0xFFFFFF32
-        self.CMD_CRC                    = 0xFFFFFF03
-        self.CMD_DIAL                   = 0xFFFFFF2D
-        self.CMD_DLSTART                = 0xFFFFFF00
-        self.CMD_EXECUTE                = 0xFFFFFF07
-        self.CMD_FGCOLOR                = 0xFFFFFF0A
-        self.CMD_GAUGE                  = 0xFFFFFF13
-        self.CMD_GETMATRIX              = 0xFFFFFF33
-        self.CMD_GETPOINT               = 0xFFFFFF08
-        self.CMD_GETPROPS               = 0xFFFFFF25
-        self.CMD_GETPTR                 = 0xFFFFFF23
-        self.CMD_GRADCOLOR              = 0xFFFFFF34
-        self.CMD_GRADIENT               = 0xFFFFFF0B
-        self.CMD_HAMMERAUX              = 0xFFFFFF04
-        self.CMD_IDCT                   = 0xFFFFFF06
-        self.CMD_INFLATE                = 0xFFFFFF22
-        self.CMD_INTERRUPT              = 0xFFFFFF02
-        self.CMD_KEYS                   = 0xFFFFFF0E
-        self.CMD_LOADIDENTITY           = 0xFFFFFF26
-        self.CMD_LOADIMAGE              = 0xFFFFFF24
-        self.CMD_LOGO                   = 0xFFFFFF31
-        self.CMD_MARCH                  = 0xFFFFFF05
-        self.CMD_MEMCPY                 = 0xFFFFFF1D
-        self.CMD_MEMCRC                 = 0xFFFFFF18
-        self.CMD_MEMSET                 = 0xFFFFFF1B
-        self.CMD_MEMWRITE               = 0xFFFFFF1A
-        self.CMD_MEMZERO                = 0xFFFFFF1C
-        self.CMD_NUMBER                 = 0xFFFFFF2E
-        self.CMD_PROGRESS               = 0xFFFFFF0F
-        self.CMD_REGREAD                = 0xFFFFFF19
-        self.CMD_ROTATE                 = 0xFFFFFF29
-        self.CMD_SCALE                  = 0xFFFFFF28
-        self.CMD_SCREENSAVER            = 0xFFFFFF2F
-        self.CMD_SCROLLBAR              = 0xFFFFFF11
-        self.CMD_SETFONT                = 0xFFFFFF2B
-        self.CMD_SETMATRIX              = 0xFFFFFF2A
-        self.CMD_SKETCH                 = 0xFFFFFF30
-        self.CMD_SLIDER                 = 0xFFFFFF10
-        self.CMD_SNAPSHOT               = 0xFFFFFF1F
-        self.CMD_SPINNER                = 0xFFFFFF16
-        self.CMD_STOP                   = 0xFFFFFF17
-        self.CMD_SWAP                   = 0xFFFFFF01
-        self.CMD_TEXT                   = 0xFFFFFF0C
-        self.CMD_TOGGLE                 = 0xFFFFFF12
-        self.CMD_TOUCH_TRANSFORM        = 0xFFFFFF20
-        self.CMD_TRACK                  = 0xFFFFFF2C
-        self.CMD_TRANSLATE              = 0xFFFFFF27
-
         # Register definitions
         self.REG_ID                     = 0x102400
         self.REG_FRAMES                 = 0x102404
